rename internal "client" package to "external"
