+++ src/external/index.ts
@@ -13,5 +13,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+export { HTTPError } from './error';
 export { Fulcio } from './fulcio';
 export { Rekor } from './rekor';
